requirements.txt
aiogram>=3.0.0
setuptools>=65.5.1
python-dotenv==1.0.1
